requirements.txt
pymongo
python-dotenv
fastapi
uvicorn
motor
requests
pillow
python-multipart
hashlib-additional
openai
